backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Devin AI <158243242+devin-ai-integration[bot]@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.135.3"}
sqlalchemy = "^2.0.49"
alembic = "^1.18.4"
pandas = "^3.0.2"
lxml = "^6.0.2"
openpyxl = "^3.1.5"
python-multipart = "^0.0.24"
aiofiles = "^25.1.0"
httpx = "^0.28.1"
python-dotenv = "^1.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
